+++ src/hash_set_sequential.h
@@ -31,9 +31,9 @@ class HashSetSequential : public HashSetBase<T> {
   bool Remove(T elem) final {
     bucket_t<T>& bucket = GetBucket(elem);
 
-    for (long i = 0; i < bucket.size(); i++) {
+    for (size_t i = 0; i < bucket.size(); i++) {
       if (bucket[i] == elem) {
-        bucket.erase(bucket.begin() + i);
+        bucket.erase(bucket.begin() + static_cast<long>(i));
         return true;
       }
     }
